requirements.txt
django==5.2.7
djangorestframework==3.16.0
drf-spectacular==0.27.2
djangorestframework-simplejwt==5.5.1
psycopg2-binary
djangorestframework-simplejwt[crypto]
django-minio-storage
python-decouple
minio
